fix: final review fixes for 3.2.1 migration (query params, path idiom, dead exports)
- add missing required/optional query params to get_stats_nodes_users_usage
  and get_users_stream (both were 400ing/silently dropping filters)
- unify path-parameter idiom across all controllers: snake_case kwarg +
  Path(alias="<specName>"), replacing the camelCase-kwarg holdouts
- delete dead model exports left over from 2.8.x (PubKeyData/PubKeyResponseDto,
  GetUserUsageByRangeResponseDto, GetNodesUsageByRangeResponseDto,
  GetNodesRealtimeUsageResponseDto, GetNodeUserUsageByRangeResponseDto,
  LastConnectedNodeDto) and their __init__ exports; rename GetPubKeyResponseDto
  to GetNodeSecretKeyResponseDto to match what it actually carries
- fix subscription_request_history test: assert current model fields instead
  of a deleted one, narrow the swallowing except to ApiError
- node_plugins.py: user_id float -> int (two occurrences); drop unused
  GetInternalSquadUsageResponseDto import from bandwidthstats.py models
- add test_query_params_match (alias-aware) to spec conformance suite, with
  a triaged allowlist for gaps found but out of scope for this wave

+++ remnapy/models/node_plugins.py
@@ -57,7 +57,7 @@ class TorrentBlockerReportRecordDto(BaseModel):
     model_config = ConfigDict(populate_by_name=True)
 
     id: float
-    user_id: float = Field(alias="userId")
+    user_id: int = Field(alias="userId")
     node_id: float = Field(alias="nodeId")
     user: TorrentBlockerUserDto
     node: TorrentBlockerNodeDto
@@ -90,7 +90,7 @@ class TorrentBlockerStatsDto(BaseModel):
 class TorrentBlockerTopUserDto(BaseModel):
     model_config = ConfigDict(populate_by_name=True)
 
-    user_id: float = Field(alias="userId")
+    user_id: int = Field(alias="userId")
     color: str
     username: str
     total: float
